requirements.txt
google-generativeai==0.8.3
pandas==2.2.3
numpy==2.1.1
streamlit==1.39.0
sqlite3
matplotlib==3.9.2
plotly==5.24.1
python-dotenv==1.0.1
scikit-learn==1.5.2
